pyproject.toml
[tool.poetry]
name = "canary-islands-professors-designations"
version = "0.1.0"
description = ""
authors = ["Adrian Abreu <adrianabreugonzalez@outlook.com>"]
readme = "README.md"
packages = [{include = "canary_islands_professors_designations"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
pypdf2 = "^2.11.1"
python-telegram-bot = {version = "^20.0a4", allow-prereleases = true}


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
